backend/requirements.txt
flask
python-dotenv
Flask-SQLAlchemy
Werkzeug
Flask-JWT-Extended
flask-cors
Flask-Mail
gunicorn
psycopg2
psycopg2-binary
